+++ src/de.secucard.connect/client-node-environment.js
@@ -40,6 +40,7 @@ export const ClientNodeEnvironment = {
         General.AccountService,
         General.AccountDeviceService,
         General.ContactService,
+        General.ContractService,
         General.DeliveryAddressService,
         General.DeviceService,
         General.FileAccessService,
@@ -83,6 +84,7 @@ export const ClientNodeEnvironment = {
         Prepaid.ContractService,
         Prepaid.ItemGroupService,
         Prepaid.ItemService,
+        Prepaid.ReportService,
         Prepaid.SaleService,
         Prepaid.StockService,
 
@@ -130,6 +132,7 @@ export const ServiceMap = {
         Accounts: General.AccountService.Uid,
         AccountDevices: General.AccountDeviceService.Uid,
         Contacts: General.ContactService.Uid,
+        Contracts: General.ContractService.Uid,
         DeliveryAddresses: General.DeliveryAddressService.Uid,
         Devices: General.DeviceService.Uid,
         FileAccesses: General.FileAccessService.Uid,
@@ -176,6 +179,7 @@ export const ServiceMap = {
         Contracts: Prepaid.ContractService.Uid,
         ItemGroups: Prepaid.ItemGroupService.Uid,
         Items: Prepaid.ItemService.Uid,
+        Reports: Prepaid.ReportService.Uid,
         Sales: Prepaid.SaleService.Uid,
         Stocks: Prepaid.StockService.Uid
     },