+++ src/de.secucard.connect/product/prepaid/prepaid.js
@@ -14,10 +14,12 @@ import {ItemGroupService} from './item-group-service';
 import {ItemService} from './item-service';
 import {SaleService} from './sale-service';
 import {StockService} from './stock-service';
+import {ReportService} from './report-service';
 
 export const Prepaid = {};
 Prepaid.ContractService = ContractService;
 Prepaid.ItemGroupService = ItemGroupService;
 Prepaid.ItemService = ItemService;
+Prepaid.ReportService = ReportService;
 Prepaid.SaleService = SaleService;
 Prepaid.StockService = StockService;