+++ src/de.secucard.connect/product/prepaid/report-service.js
@@ -9,6 +9,22 @@
  See the License for the specific language governing permissions and
  limitations under the License.
  */
-export const Version = {
-    "name": "0.5.4"
-};
+import {ProductService} from '../product-service'
+
+export class ReportService extends ProductService {
+
+    constructor() {
+        super()
+    }
+
+    getEndpoint() {
+        return ['prepaid', 'reports'];
+    }
+
+    getEventTargets() {
+        return [];
+    }
+
+}
+
+ReportService.Uid = (['prepaid', 'reports']).join('.');