+++ Changelog.md
@@ -1,3 +1,13 @@
+<a name="0.5.4"></a>
+## 0.5.4 (2019-01-16)
+
+* added endpoint for general contracts
+
+<a name="0.5.3"></a>
+## 0.5.3 (2018-11-29)
+
+* added endpoint for prepaid reports
+
 <a name="0.5.2"></a>
 ## 0.5.2 (2018-08-02)